refactor: split render runtime helpers into modules
